+++ nest-backend/src/auth/auth.controller.ts
@@ -1,4 +1,35 @@
-import { Controller } from '@nestjs/common';
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Param,
+  Delete,
+  Put,
+  Res,
+} from '@nestjs/common';
+import { CreateUserDto } from 'src/users/dto/create-user.dto';
+import { Response } from 'express';
+import { AuthService } from './auth.service';
 
 @Controller('auth')
-export class AuthController {}
+export class AuthController {
+  constructor(private readonly authService: AuthService) {}
+  @Post('signIn')
+  async signIn(
+    @Body() createUserDto: CreateUserDto,
+    @Res({ passthrough: true }) res: Response,
+  ): Promise<any> {
+    const tokens = await this.authService.signIn(createUserDto);
+    res.cookie('accessToken', tokens.accessToken, { httpOnly: true });
+    res.cookie('refreshToken', tokens.refreshToken, { httpOnly: true });
+    return res.send(tokens);
+  }
+
+  @Delete('signOut')
+  async signOut(@Res({ passthrough: true }) res: Response) {
+    res.clearCookie('accessToken');
+    res.clearCookie('refreshToken');
+    return { message: 'Sign Out Successfully' };
+  }
+}
